third_party/blink/web_tests/fast/frames/sandboxed-iframe-navigation-allowed-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE ERROR: Failed to load '' as a plugin, because the frame into which the plugin is loading is sandboxed.
This test verifies that a sandboxed iframe CAN navigate both itself and a child in the frame tree. It also verifies that the sandbox attribute remains intact after a frame has been navigated, and that sandbox attributes are propagated through the frame hierarchy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS (self.sandboxedFrame.frameWithPlugin.document.getElementById('plugin').postMessage) == undefined is true
PASS (self.frameWithPlugin.document.getElementById('plugin').postMessage) != undefined is true
PASS document.getElementById('sandboxedFrameId').sandbox is ['allow-scripts', 'allow-same-origin']
PASS document.getElementById('sandboxedFrameId').getAttribute('sandbox') is 'allow-scripts'
PASS successfullyParsed is true

TEST COMPLETE

 
